Switch all patches over to HarmonyLib
